Bosstag changes and bugfixes (#3788)
* bosstag changes for star point mobs

* what's updog

* boss_coordinatedattack and starfall

* one more param

* random param fixes

* bugfixes for some params

* slight adjustment to boss_projectile glowing

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/bosses/bosses/MobRisingBoss.java
@@ -23,7 +23,7 @@ public class MobRisingBoss extends BossAbilityGroup {
 		public int DELAY = 80;
 		public int COOLDOWN = 160;
 		public int DURATION = 80;
-		public int RANGE = 20;
+		public double RANGE = 20;
 		public boolean CAN_BE_STOPPED = false;
 		public boolean CAN_MOVE = false;
 		public boolean SINGLE_TARGET = false;
@@ -59,14 +59,14 @@ public class MobRisingBoss extends BossAbilityGroup {
 				},
 				() -> {
 					if (ZoneUtils.hasZoneProperty(boss.getLocation(), ZoneUtils.ZoneProperty.RESIST_5)
-						    && !ZoneUtils.hasZoneProperty(boss.getLocation(), ZoneUtils.ZoneProperty.BLITZ)) {
+						&& !ZoneUtils.hasZoneProperty(boss.getLocation(), ZoneUtils.ZoneProperty.BLITZ)) {
 						return new ArrayList<>();
 					}
 					return p.TARGETS.getTargetsLocationList(boss);
 				},
 				(Location loc, int times) -> {
 					if (ZoneUtils.hasZoneProperty(loc, ZoneUtils.ZoneProperty.RESIST_5)
-						    && !ZoneUtils.hasZoneProperty(boss.getLocation(), ZoneUtils.ZoneProperty.BLITZ)) {
+						&& !ZoneUtils.hasZoneProperty(boss.getLocation(), ZoneUtils.ZoneProperty.BLITZ)) {
 						return null;
 					}
 					return p.MOB_POOL.spawn(loc);
